+++ gpflow/inducing_variables/__init__.py
@@ -1,4 +1,4 @@
 from .inducing_variables import InducingVariables, InducingPoints, Multiscale
-from .convolutional_patch import InducingPatches
+from .inducing_patch import InducingPatches
 from .mo_inducing_variables import (MultioutputInducingVariables, FallbackSharedIndependentInducingVariables, FallbackSeparateIndependentInducingVariables,
                                     SharedIndependentInducingVariables, SeparateIndependentInducingVariables)
