+++ gpflow/inducing_variables/inducing_patch.py
